Narrow generated Builder capability
