+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/PostTreePhase.java
@@ -35,7 +35,7 @@ public class PostTreePhase extends BuilderVisitingPhaseAction {
     }
 
     @Override
-    protected void doVisit(@NotNull String path, @NotNull KlumBuilder<?> builder, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
+    protected void doVisit(@NotNull String path, @NotNull InternalKlumBuilder<?> builder, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
         LifecycleHelper.executeLifecycleMethods(builder, PostTree.class);
     }
