+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/layer3/AutoLinkPhase.java
@@ -25,7 +25,7 @@ package com.blackbuild.klum.ast.util.layer3;
 
 import com.blackbuild.klum.ast.process.DefaultKlumPhase;
 import com.blackbuild.klum.ast.process.BuilderVisitingPhaseAction;
-import com.blackbuild.klum.ast.util.KlumBuilder;
+import com.blackbuild.klum.ast.util.InternalKlumBuilder;
 import com.blackbuild.klum.ast.util.LifecycleHelper;
 import com.blackbuild.klum.ast.util.layer3.annotations.AutoLink;
 import com.blackbuild.klum.ast.util.layer3.annotations.LinkTo;
@@ -39,7 +39,7 @@ public class AutoLinkPhase extends BuilderVisitingPhaseAction {
     }
 
     @Override
-    protected void doVisit(@NotNull String path, @NotNull KlumBuilder<?> element, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
+    protected void doVisit(@NotNull String path, @NotNull InternalKlumBuilder<?> element, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
         ClusterModel.getFieldsAnnotatedWith(element, LinkTo.class)
                 .entrySet()
                 .stream()
