+++ klum-ast/src/main/java/com/blackbuild/groovy/configdsl/transform/ast/ProxyMethodBuilder.java
@@ -28,7 +28,7 @@ import com.blackbuild.annodocimal.ast.formatting.AnnoDocUtil;
 import com.blackbuild.annodocimal.ast.formatting.DocBuilder;
 import com.blackbuild.annodocimal.ast.formatting.DocText;
 import com.blackbuild.klum.ast.util.FactoryHelper;
-import com.blackbuild.klum.ast.util.KlumBuilder;
+import com.blackbuild.klum.ast.util.InternalKlumBuilder;
 import com.blackbuild.klum.ast.util.TemplateManager;
 import com.blackbuild.klum.ast.util.reflect.AstReflectionBridge;
 import groovy.lang.Closure;
@@ -55,7 +55,7 @@ public final class ProxyMethodBuilder extends AbstractMethodBuilder<ProxyMethodB
 
     private static final ClassNode FACTORY_HELPER_TYPE = make(FactoryHelper.class);
     private static final ClassNode TEMPLATE_MANAGER_TYPE = make(TemplateManager.class);
-    private static final ClassNode KLUM_BUILDER_TYPE = make(KlumBuilder.class);
+    private static final ClassNode KLUM_BUILDER_TYPE = make(InternalKlumBuilder.class);
 
     private final String proxyMethodName;
     private final Expression proxyTarget;
